complete building on Windows

+++ src/nbt.cc
@@ -342,7 +342,8 @@ namespace mcpe_viz {
       // todo - how to report invalid in geojson?
       // if ( valid ) {
       std::ostringstream str;
-      if ( std::isnan(x) || std::isnan(y) ) {
+      // hack: fix error C2668 'fpclassify': ambiguous call to overloaded function
+      if ( std::isnan(double(x)) || std::isnan(double(y)) ) {
         // we don't put out anything because "NaN" is not valid JSON
         str << "";
       } else {
@@ -390,7 +391,8 @@ namespace mcpe_viz {
       // todo - how to report invalid in geojson?
       // if ( valid ) {
       std::ostringstream str;
-      if ( std::isnan(x) || std::isnan(y) || std::isnan(z) ) {
+      // hack: fix error C2668 'fpclassify': ambiguous call to overloaded function unittest
+      if ( std::isnan(double(x)) || std::isnan(double(y)) || std::isnan(double(z)) ) {
         // we don't put out anything because "NaN" is not valid JSON
         str << "";
       } else {
